msm: audio: qdsp6v2: Fix in_call voice recording issue
During in_call voice recording with both rx+tx options
msm_enable_incall_recording fucntion is sending array of port ids
which are declared as int array. But the adm_matrix_map function
is taking port ids as unsigned short *, because of this datatype
mismatch the second port id is read as 0 instead of 32771.
Fix this issue by maintaining the port ids in unsigned int array
and adm_matrix_map proto type is changed to take port ids as
unsigned int array.

CRs-fixed: 292705
(cherry picked from commit I5ef6968289673b656d5901183c7b468b6918fbd7)

+++ arch/arm/mach-msm/qdsp6v2/q6adm.c
@@ -298,7 +298,7 @@ fail_cmd:
 }
 
 int adm_matrix_map(int session_id, int path, int num_copps,
-			unsigned short *port_id, int copp_id)
+			unsigned int *port_id, int copp_id)
 {
 	struct adm_routings_command	route;
 	int ret = 0, i = 0;
